Removing nullability of 'idKey' and defaulting to database.driver.idKey, or simply "id", in case no database is available.
This avoids cluttering the code with nullability checks, but has the drawback of not being able to identify if the default idKey is intended to override the database's idKey.

Also makes an entity's idKey return different values, depending on which database is assigned to Self, currently, which seems to be a change in previous behavior.