Use valid UUID for insert-and-fix
Postgres – unlike MySQL – actually validates the UUIDs sent as values, so the tests of postgresql-driver were silently broken by the Fluent 2.0.1 release. This commit fixes the test by using a valid UUID instead of just `"asdf"`.

I noticed that PG returns UUID fields as all lowercase, so I changed the comparison to be case insensitive. (It shouldn't be a problem as we're talking about the string representation of an otherwise fixed binary entity and the hex numbers still match.)